--- src/main/java/site/dittotrip/ditto_trip/hashtag/domain/entity/CategoryHashtag.java
@@ -1,27 +1,0 @@
-package site.dittotrip.ditto_trip.hashtag.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import site.dittotrip.ditto_trip.category.domain.Category;
-
-import static jakarta.persistence.FetchType.*;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class CategoryHashtag {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_hashtag_id")
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "hashtag_id")
-    private Hashtag hashtag;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "category_id")
-    private Category category;
-
-}
