--- src/main/java/site/dittotrip/ditto_trip/hashtag/domain/entity/SpotHashtag.java
@@ -1,27 +1,0 @@
-package site.dittotrip.ditto_trip.hashtag.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import site.dittotrip.ditto_trip.spot.domain.Spot;
-
-import static jakarta.persistence.FetchType.LAZY;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class SpotHashtag {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "spot_hashtag_id")
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "hashtag_id")
-    private Hashtag hashtag;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "spot_id")
-    private Spot spot;
-
-}
